+++ gui/model/ratamodel.cpp
@@ -60,7 +60,7 @@ QVariant RataModel::headerData(int section, Qt::Orientation orientation, int rol
 
 QModelIndex RataModel::parent(const QModelIndex &child) const
 {
-    if (child.internalId() != -1) {
+    if (static_cast<int>(child.internalId()) != -1) {
         return createIndex(m_sarjat.indexOf(static_cast<Sarja*>(child.internalPointer())), 0, -1);
     }
 
@@ -86,7 +86,7 @@ QVariant RataModel::data(const QModelIndex &index, int role) const
         return QVariant();
     }
 
-    if (index.internalId() == -1) {
+    if (static_cast<int>(index.internalId()) == -1) {
         Sarja* s = m_sarjat.at(index.row());
 
         switch (index.column()) {
@@ -127,7 +127,7 @@ bool RataModel::setData(const QModelIndex &index, const QVariant &value, int rol
     bool res = false;
 
     // Rata editointi
-    if (index.internalId() == -1) {
+    if (static_cast<int>(index.internalId()) == -1) {
         Sarja *s = m_sarjat.at(index.row());
 
         switch (index.column()) {
